UUID (#322)
# Description

This adds a custom UUID to GoMud. It is 16 bytes in size, and reserves 56 bits for future expansion.

The first 1 byte is the version of the UUID. Version 0 is reserved for `nil` UUID's. The maximum version will be 15 before having to make a more dramatic change.
Incrementing versions allows us to have custom formats if needed.

`0` or empty (``) UUID's will be parsed into nil UUID's. When stringified, they will be `0` strings. This will remain consistent across versions.

The next 7 bytes are unix microseconds since a custom epoch (January 1, 2025), plus a 1 byte sequence counter for microsecond collisions.

It allows for up to 256 custom "types", stored as 1 byte. This will likely be used as a quick category lookup, such as checking whether it is a UUID for an item... but isn't required to be used. (Will default zero).

Since it is a byte array, it is a `comparable` type and can be used as a map key etc.

It serializes to a hexadecimal hyphen separated string (see `String Format of UUID`)

## String Format of UUID

 Format is:

`<version:1 hex digit>-<timestamp:13 hex digits><sequence:2 hex digits>-<type:2 hex digits>-<unused:14 hex digits>`

Example "nil" UUID:
**UUID:** `""`
**UUID:** `"0"`
**UUID:** `"0..."`

Example non nil UUID:
**UUID:** `"1-00877c496c2bbdc-b0-00000000000000"`
**Version:** `1`
**Time:** `2025-04-18 11:14:52.345019 -0700 PDT`
**Sequence:** `220`
**Type:** `176`

## Changes

- Added `uuid` package.
- Added `UUID` field to items, using for comparisons.
- Added UUID's to items. These are currently excluded from yaml when Marshaling. Later they may be left in. This means they are generated anew when loaded back into the world.
  - UUID's are now used in item comparisons.
- Adjusted buff to use `OnStartWaiting` instead of `OnStartEvent` to save on writing unnecessary data.