CONTRIBUTING.md and LICENSE.md changes
# Description

This adds a CONTRIBUTING.md, instructions on how to contribute code to GoMud.
Additionally it moves the LICENSE.md out of the `.github` folder, where it is not being recognized, and into the root of the repo, where it should be.

## Changes

Provide a bullet point list of noteworthy changes in this Pull Request:

- `.github/CONTRIBUTING.md`
- `.github/LICENSE.md` => `LICENSE.md`
- Linked to `CONTRIBUTING.md` from `README.md` ( Contributor Guide section )

## Issues

https://github.com/GoMudEngine/GoMud/issues/330