Adding a prompt for kicking user when double-logging in (#345)
# Description

When a user is already logged in, and tries to log in again, it denies them and disconnects.

However, there may be situations where they are not really logged in, and something is maintaining the connection, preventing them from logging in until an inactivity timeout kicks the "almost zombie" user.

This adds a prompt to the login process allowing a second login to optionally kick the first login, and replace them.

## Changes

- Added a new prompt: "User is already connected. Kick them?"
- When `no` is selected, they are denied login as it is currently.
- When `yes` is selected, the currently connected user is kicked, and the new login replaces them.
- Removed some unused code/functions