Mob pathfinding (#341)
# Description

This adds initial implementation of pathfinding for mobs. 
A `pathto` command is added, which accepts multiple room id's, or the keyword `home` (to go back to their spawned-from room) to initiate a path for a mob. This allows testing via the admin `command` command, and also initiating via scripting.

## Changes

Provide a bullet point list of noteworthy changes in this Pull Request:

- Paths are followed when mobs are idle
- `pathto` command added.
  - Example: `pathto 300 1 home` to create/follow a path to roomId 300, then 1, then back to home room.
  - If rooms cannot be fully connected, `pathto` fails without a path.
- Removed old "room stack" tracking used to return home when mobs wander too far. It was buggy anyways. Replaced with `pathto home`
- Added some randomly selected paths for frostfang guards to patrol.
- Scripting functions to check path status