Changing handling on all-modules.go and the 'go generate' feature (#355)
# Description

When forking the repository, and running "go generate" modules/all-modules.go is created or updated, but since it was previously tracking in the repository it's then added to unstaged files for commit. This is not optimal, as it could potentially mean someone would push their custom all-modules.go file to the main repository.

It's already in .gitignore, but since it was already tracked this is not enough to have ti not appear as a changed file when running "go generate".

## Changes

- Added a default modules/all-modules.go to the repo
- Removed all-modules.go from .gitignore
- Removed references to 'go generate' in the Makefile
- Removed cmd/generate/_all-modules.go
- Added proper sorting of modules in modules-import.go