Fixing mapper to generate maps for all rooms, including disjointed ones. (#358)
# Description

The mapper generates a map of rooms in a zone, starting from the Root Room ID.

However, if the zone contains other rooms that are disconnect from the main group of rooms, or are reached via an unconventional exit (doesn't have a coordinate direction it exits to), they are being left out.

While still disconnected, this change allows those areas to have their own sub-maps that are accessed based on roomId now instead of a blanket zone name.


## Changes

- mapper now has a separate cache that maps roomId's to their "zone cache"
- Simplified the pre-cache logic
- Renamed `GetZoneMapper()` => `GetMapper()`